+++ src/files.rs
@@ -1,7 +1,7 @@
 // files.rs
 
 // *************************************************************************
-// * Copyright (C) 2019-2020 Daniel Mueller (deso@posteo.net)              *
+// * Copyright (C) 2019-2022 Daniel Mueller (deso@posteo.net)              *
 // *                                                                       *
 // * This program is free software: you can redistribute it and/or modify  *
 // * it under the terms of the GNU General Public License as published by  *
@@ -143,8 +143,8 @@ pub mod test {
 
   use crate::keys::FromPem;
 
-  use ssh_agent::proto::private_key::PrivateKey;
-  use ssh_agent::proto::public_key::PublicKey;
+  use ssh_agent_lib::proto::private_key::PrivateKey;
+  use ssh_agent_lib::proto::public_key::PublicKey;
 
 
   /// Load a private key from a plain text file. This function is for