+++ src/keys.rs
@@ -1,7 +1,7 @@
 // keys.rs
 
 // *************************************************************************
-// * Copyright (C) 2019-2020 Daniel Mueller (deso@posteo.net)              *
+// * Copyright (C) 2019-2022 Daniel Mueller (deso@posteo.net)              *
 // *                                                                       *
 // * This program is free software: you can redistribute it and/or modify  *
 // * it under the terms of the GNU General Public License as published by  *
@@ -24,10 +24,10 @@ use anyhow::anyhow;
 use anyhow::Context as _;
 use anyhow::Result;
 
-use ssh_agent::proto::private_key::Ed25519PrivateKey;
-use ssh_agent::proto::private_key::PrivateKey;
-use ssh_agent::proto::public_key::Ed25519PublicKey;
-use ssh_agent::proto::public_key::PublicKey;
+use ssh_agent_lib::proto::private_key::Ed25519PrivateKey;
+use ssh_agent_lib::proto::private_key::PrivateKey;
+use ssh_agent_lib::proto::public_key::Ed25519PublicKey;
+use ssh_agent_lib::proto::public_key::PublicKey;
 
 use ssh_keys::openssh::parse_private_key;
 use ssh_keys::openssh::parse_public_key;